requirements.txt
python-dotenv==0.17.1
pprintpp==0.4.0
numpy==1.22.3
pytest==6.2.4
ta==0.7.0
cbpro==1.1.4